compiler warnings in solarpilot and nlopt

+++ solarpilot/AutoPilot_API.cpp
@@ -1582,7 +1582,7 @@ bool AutoPilot::OptimizeAuto(vector<double*> &optvars, vector<double> &upper_ran
         start.at(i) = *optvars.at(i);
 
     //Check feasibility
-    unsigned int iht = std::find(names->begin(), names->end(), "receiver.0.rec_height") - names->begin();
+    unsigned int iht = (unsigned int)(std::find(names->begin(), names->end(), "receiver.0.rec_height") - names->begin());
     if( iht < names->size() )
     {
         double *xtemp = new double[ optvars.size() ];
